requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AWS SDK
boto3==1.34.25
botocore==1.34.25

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.92.2
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
